refactor(connect): move blockchain method definitions to corresponding folder
