+++ packages/connect-common/src/types/api/blockchain/blockchainDisconnect.ts
@@ -1,4 +1,4 @@
-import type { CommonParamsWithCoin, Response } from '../params';
+import type { CommonParamsWithCoin, Response } from '../../params';
 
 export declare function blockchainDisconnect(
     params: CommonParamsWithCoin,
