+++ packages/connect-common/src/types/api/blockchain/blockchainGetContractInfo.ts
@@ -1,7 +1,7 @@
 import type { BlockchainLinkResponse } from '@trezor/blockchain-link';
 import { type ContractInfoParams } from '@trezor/blockchain-link-types/src/blockbook';
 
-import type { CommonParamsWithCoin, Response } from '../params';
+import type { CommonParamsWithCoin, Response } from '../../params';
 
 export declare function blockchainGetContractInfo(
     params: CommonParamsWithCoin & ContractInfoParams,
