+++ packages/connect-common/src/types/api/blockchain/blockchainGetFiatRatesForTimestamps.ts
@@ -1,6 +1,6 @@
 import type { BlockchainLinkParams, BlockchainLinkResponse } from '@trezor/blockchain-link';
 
-import type { CommonParamsWithCoin, Response } from '../params';
+import type { CommonParamsWithCoin, Response } from '../../params';
 
 export declare function blockchainGetFiatRatesForTimestamps(
     params: CommonParamsWithCoin & BlockchainLinkParams<'getFiatRatesForTimestamps'>,
