+++ packages/connect-common/src/types/api/blockchain/blockchainGetTransactions.ts
@@ -1,6 +1,6 @@
 import type { Transaction } from '@trezor/blockchain-link';
 
-import type { CommonParamsWithCoin, Response } from '../params';
+import type { CommonParamsWithCoin, Response } from '../../params';
 
 export type BlockchainGetTransactions = CommonParamsWithCoin & {
     txs: string[];
