+++ packages/connect-common/src/types/api/blockchain/blockchainSetCustomBackend.ts
@@ -1,5 +1,5 @@
-import type { BlockchainLink } from '../coinInfo';
-import type { CommonParams, Response } from '../params';
+import type { BlockchainLink } from '../../coinInfo';
+import type { CommonParams, Response } from '../../params';
 
 export type BlockchainSetCustomBackend = CommonParams & {
     coin: string;
