+++ packages/connect-common/src/types/api/blockchain/blockchainSubscribeFiatRates.ts
@@ -1,6 +1,6 @@
 import type { BlockchainLinkResponse } from '@trezor/blockchain-link';
 
-import type { CommonParamsWithCoin, Response } from '../params';
+import type { CommonParamsWithCoin, Response } from '../../params';
 
 export type BlockchainSubscribeFiatRates = CommonParamsWithCoin & {
     currency?: string;
