+++ packages/connect-common/src/types/api/blockchain/blockchainUnsubscribeFiatRates.ts
@@ -1,7 +1,7 @@
 import type { BlockchainLinkResponse } from '@trezor/blockchain-link';
 
-import type { Response } from '../params';
 import type { BlockchainSubscribeFiatRates } from './blockchainSubscribeFiatRates';
+import type { Response } from '../../params';
 
 export declare function blockchainUnsubscribeFiatRates(
     params: BlockchainSubscribeFiatRates,
