+++ packages/connect-common/src/types/api/blockchain/pushTransaction.ts
@@ -5,7 +5,7 @@
 import type { Static } from '@trezor/schema-utils';
 import { Type } from '@trezor/schema-utils';
 
-import type { Params, Response } from '../params';
+import type { Params, Response } from '../../params';
 
 export type PushTransaction = Static<typeof PushTransaction>;
 export const PushTransaction = Type.Object({
